feat: added ability for users to set account-specific file dimensions (#402)
* feat: added ability for users to set account-specific file dimensions

* chore: set console to open only when asked
