+++ apps/postybirb-ui/src/components/submissions/submission-edit-form/submission-file-manager/submission-file-card/submission-file-card.tsx
@@ -3,13 +3,14 @@ import {
   FileSubmissionMetadata,
   FileType,
   ISubmissionFileDto,
+  NULL_ACCOUNT_ID,
 } from '@postybirb/types';
 import { getFileType } from '@postybirb/utils/file-type';
 import { IconGripVertical } from '@tabler/icons-react';
 import { SubmissionDto } from '../../../../../models/dtos/submission.dto';
 import { FileCardDeleteAction } from './file-card-delete-action';
 import { FileCardFileActions } from './file-card-file-actions';
-import { FileMetadataManager } from './file-metadata-manager';
+import { FileMetadataManager } from './file-metadata-manager/file-metadata-manager';
 import { FileTextAlt } from './file-text-alt';
 import { FileValidations } from './file-validations';
 
@@ -42,44 +43,47 @@ export function SubmissionFileCard({
       className={DRAGGABLE_SUBMISSION_FILE_CLASS_NAME}
     >
       {draggable && (
-        <Box 
-          style={{ 
-            position: 'absolute', 
-            left: 8, 
-            top: '50%', 
+        <Box
+          style={{
+            position: 'absolute',
+            left: 8,
+            top: '50%',
             transform: 'translateY(-50%)',
             opacity: 0.5,
-            cursor: 'grab'
+            cursor: 'grab',
           }}
         >
           <IconGripVertical size={18} />
         </Box>
       )}
-      
+
       <Flex gap="xl" align="flex-start" key="card-file-previewer">
-        <Box 
-          style={{ 
+        <Box
+          style={{
             // eslint-disable-next-line lingui/no-unlocalized-strings
-            flex: '0 0 auto', 
-            padding: '4px', 
-            borderRadius: '8px', 
-            background: 'var(--mantine-color-dark-6)'
+            flex: '0 0 auto',
+            padding: '4px',
+            borderRadius: '8px',
+            background: 'var(--mantine-color-dark-6)',
           }}
         >
           <FileCardFileActions file={file} submissionId={submission.id} />
         </Box>
-        
+
         <Box style={{ flex: 1 }}>
           <FileValidations submission={submission} file={file} />
-          
+
           <Divider my="sm" variant="dashed" />
-          
+
           <FileMetadataManager
             submissionId={submission.id}
             file={file}
             metadata={metadata}
+            accounts={submission.options
+              .map((option) => option.account)
+              .filter((account) => account.id !== NULL_ACCOUNT_ID)}
           />
-          
+
           {getFileType(file.fileName) === FileType.TEXT ? (
             <Box mt="md" style={{ position: 'relative' }}>
               <Divider my="sm" variant="dashed" />
@@ -87,7 +91,7 @@ export function SubmissionFileCard({
             </Box>
           ) : null}
         </Box>
-        
+
         <Box style={{ alignSelf: 'flex-start' }}>
           <FileCardDeleteAction
             submissionId={submission.id}
