+++ apps/postybirb/src/app/app.ts
@@ -123,7 +123,7 @@ export default class PostyBirb {
       PostyBirb.mainWindow.show();
 
       // if (PostyBirb.isDevelopmentMode()) {
-      PostyBirb.mainWindow.webContents.openDevTools({ mode: 'detach' });
+      // PostyBirb.mainWindow.webContents.openDevTools({ mode: 'detach' });
       // }
     });
